[zither] Introduce 'output namespace' knob
"Output namespace" is defined to be the namespace/layout under which
backend outputs are generated within the specified output directory. By
default, this is backend-specific. The value can determine the name of
the resulting library/package/crate/etc. (when a function of source
layout), as well as the 'include' namespace of the headers generated by
a C family backend.

For now, we just add this support for the C family backends (though this
support can be trivially extended to the other data layout backends when
a use-case presents itself). An explicit override will permit us to
remove the header guard rewrite logic currently hard-coded for the "zbi"
and "zx" FIDL libraries.

Change-Id: I57f9d8b2904c3a18f581b48d779845316c5f2bbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828896
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>