[devicetree] Add support for aliases
The aliases node is a special node, that a devicetree::Devicetree will
keep tabs on.

From now on, PropertyDecoder keeps a pointer to the aliases node of the
devicetree, as soon as the node is found during a walk, a devicetree
stores a copy of the |Properties|, which in turn makes PropertyDecoder
reference the aliases of the devicetree.

Additionally PropertyDecoder now provides |ResolvePath| which has three
possible results:
 -> error (false): The aliases node has already been found, but the
alias in the path is not in there. This answer the question, is this
error temporal.

 -> error (true): The aliases node has not been found yet, so users may
choose to try again later after finishing a full tree walk.

 -> success(ResolvedPath): The path is either not an alias and was
resolved, or the path is an aliases, the aliases node has been found AND
the alias was within the aliases node properties.

Bug: 78675
Test: fx test -v -o devicetree-tests
Change-Id: I78bd5352e2a2b50e22af11afa2b4d3b65921ff04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826217
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>