[rust][rbe] Fix local/remote consistency check
This legacy shell-script RBE wrapper was broken for some time
in that it failed to actually compare outputs.
This gave the false impression that everything was fine.

This fix results in reporting consistent findings with the new
rustc_remote_wrapper.py.  Consistency checks are now consistent with
each other.

Bug: 125891
Change-Id: I6e29e515e48f39c9a9b1d5d6094aaedd1ae67b58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/844516
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>