[bt][obex] Define a type for OBEX request packets
The OBEX Request (e.g. Connect, Disconnect, ..) contains an OpCode and
potential headers.

Implement Encoding and Decoding for this type.

Define the HeaderSet as a collection of Headers that can be sent with an
OBEX request or response. This collection is initially simple with a
TODO to improve and enforce ordering properties defined in the OBEX
specification.

Bug: 120012
Test: fx test bt-obex-tests
Change-Id: I80a90e2fef3f7267897d0333e2ef343a83b2ceb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832208
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>