[rbe] Support remote != local command
Distinguish between remote-only and local-only commands.
Accept both, while still defaulting to remote == local commands.

Sometimes the command for remote execution is modified so that
it cannot run locally, but it is expected to produce the same result.
For example, with remote-cross-compiling, the remote request's command
might use a binary for the remote execution platform, which is not
suitable for local host execution.

Special handling for local fallback: when the remote and local commands
differ and rewrapper failed in one of the local execution modes,
retry with the local command we were given if it is different from the
remote one we gave to rewrapper.

Not applying this to racing mode because rewrapper will ignore local
failures if the remote one succeed.

--compare mode will use the correct local command when it is different
than the remote command.  This helps verify that different commands
indeed produce the same results.

When local and remote execution results differ AND the commands differ,
show the unified diff of those commands.

Bug: 105799

Change-Id: I70541aab341d38d189f3b69fdd89047d4a26f037
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845016
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Fang <fangism@google.com>