[ffx][fho] Only output any given tool once
If a tool exists in both the sdk and workspace, or the ffx binary itself
and either of those, it'll currently print twice. Use a hashset to make
sure we're not doing that and only displaying them in priority order.

Fixed: 124402
Change-Id: Ic8935aefd8bd638275798cb74bd29de5cb4cb7b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845756
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>