[sys] Rename CapabilitySourceInterface to CapabilitySource
CapabilitySource is an enum. While one could argue that
the word "Interface" makes sense for "ComponentInstanceInterface" since
that's a trait, that still doesn't follow the Rust naming convention.
The analogy is even less tenable for the CapabilitySource enum.

Is there a reasoning behind the "Interface" suffix? If not, it seems
less confusing to remove it.

Change-Id: Ic06657c78056e3c4576584da6e306015bfbd025b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845741
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>