[kernel][riscv64][phys] Update shadow-call-stack support to use gp
The compiler is changing from using s2 (x18) to using gp (x3) for
shadow-call-stack. This updates the lib/arch and phys code to work
with the newer compiler. Though this is not yet exploited, the
<lib/arch/asm.h> macro is updated to use s11 (x27) for the percpu
pointer in the kernel proper, which previously used gp.

Bug: 123949
Change-Id: Ie0f755f7f66ca1ea34fa5b14722e6f5e1f59aa05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837296
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>