[libc] Update riscv64 setjmp/longjmp for shadow-call-stack pointer in gp
The compiler is changing from using s2 (x18) to using gp (x3) for
shadow-call-stack. This updates setjmp and longjmp to work with
this.  This requires adding another word to jmp_buf because
previously s2 was saved as the shadow-call-stack pointer while gp
was not saved as it's fixed in the old ABI, but now both must be
saved because gp is the shadow-call-stack pointer and s2 is
call-saved.

This also fixes a couple of bugs in the original implementation:
 * The shadow-call-stack pointer was saved but not restored.
 * The fs0 and s11 slots overlapped, so s11 was clobbered and the
   jmp_buf was one word smaller than it should have been.

Bug: 123949
Change-Id: I1ff1aa216a349e28cb91adb65acaae75810f35d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/837298
Reviewed-by: Paul Kirth <paulkirth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>