[display][virtio] Move driver to //src/graphics/display/drivers
The //src/graphics/drivers/virtio driver is the guest-mode
driver of virtio-gpu device; it is only used for framebuffer
scanout and not for any other graphics virtualization, so
this change moves the driver to //src/graphics/display/drivers.

To better distinguish the driver from the host-mode driver
of virtio-gpu in fuchsia (located in //src/virtualization),
we rename the driver directory to "virtio-guest" to reflect
this.

Test: no-op. fx set core.x64 ; fx build ; fx qemu
Change-Id: I58cd69f3f4a9eae6ebda8df15c8b29fabc3a1c90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845576
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>