[starnix] Basic kernel panic support.
When a single process in the kernel panics, we need to tear down
the entire kernel which includes all the processes in the same job.

This is implemented as a small library that can be integration tested
without adding any new panicking paths to Starnix or depending on
particular panicking behaviors of the kernel's current implementation.

Manually confirmed that the test times out if the panic hook is
commented out from panicker.rs.

Multiply: kill_job_on_panic_tests

Change-Id: I53474868066ed65cbe488d2abad3048bd9d25896
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830879
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>