[network] Add socket benchmarks with tracing
Add variants of the existing socket benchmarks to run against NS2 with
tracing enabled, and collect traces (on the target side inside the
test logic itself using the `fuchsia.tracing.controller/Controller`
FIDL directly) to be included in the test artifacts.

Note that although these tests do not rely on any host-side machinery,
they are run as a perf test as there is no other way to include the
trace as a test artifact currently due to builders not using ffx in
infra.

Test: fx test --e2e netstack_benchmarks_test
Fixed: 109959
Change-Id: I00411f263cf6181ccdce98f291b4ef984d231b51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815080
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Tony Gong <gongt@google.com>