[kernel][vm] Writeback end updates generation count
WritebackEndLocked will remove clean zero intervals from the page list,
or clip the starting point of zero intervals. Update the generation
count to account for this page list update. Note that this bug existed
prior to the introduction of zero intervals, as we would be required to
update the generation count on updating the supply_zero_offset as well.

Bug: 125939
Run-All-Tests: true
Change-Id: I7f5c640f1a6f13a8b04f8c80e5358bc044b6b489
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/846718
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>