[netdevice] Add tests for minimum TX body len
Assert that when sending an ethernet frame via a packet socket (both
link and network sockets), both netstacks pad the outgoing frame in
accordance with the device's required size. Netstack3 currently adds
additional padding beyond what the device requires, which is apparent
for small frames. Mark those tests as failing, to be addressed soon.
Some of the added tests are failing for Netstack2 because it doesn't
correctly identify the frame type of outgoing frames to the driver, so
mark those tests failing as well.

Bug: 126190, 126192
Change-Id: I3b8cb68c986d9c1c27a788adc487058ee64b8c73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841376
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>