[kernel][diags] Clean up "dump_thread" names
Move most of the stand-alone dump_thread diagnostic routines so that
they are members of Thread, Thread::Current, or a newly introduced
ThreadDumper.

This rearrangement removes a few redundant declarations, and makes it
a bit easier to express the static locking requirements as we move to
break up the global thread lock.

Change-Id: I21bbb0ffa355a455ac3167e32c158a9e60b0658f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847296
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>