[session_manager] Add SC option for autolaunch.
This adds a structured config option to the session-manager component
indicating whether session-manager should launch the default session
as soon as it starts up.

Because developer overrides aren't yet supported, I've also added in a
hack: there's a magic file in the `data` directory, which forces
`autolaunch` to be `false`. This should be replaced with a
well-supported SC feature as soon as one is available.

To turn off autolaunch, run:

    fx shell 'mkdir -p /data/session-manager && echo "" > /data/session-manager/noautolaunch'

To re-enable it:

    fx shell 'rm /data/session-manager/noautolaunch'

I also added some integration tests.

Fixed: https://b/issues/277078620

Change-Id: I38ae973cdc9471820b61ae3b222a4bec5d0ec89e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843796
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>