[boot_test] Ensure disabled boot tests emit test metadata
Previously, we would not emit test metadata for disabled boot tests, as
the indicator for whether it was a boot test was lost (being specified
within test environment metadata); otherwise test machinery would assume
it was a normal test and would try to run the disabled test in the wrong
way. With recent metadata refactors this is no longer an issue, so we no
longer need to suppress test metadata. The main upshot of this is that
`fx run-boot-test` can now still identify and run disabled boot tests.

Fixed: 122881
Change-Id: I2372b0a5ea7374f025311f2b33ea5bc008402edc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848096
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>