[build] Always use the local pool for non-remoted actions
Always use the 'local' pool for locally-executed actions, instead of
the full (10*num_cores) concurrency that is available when goma is
enabled.

Correct the logic that sets the use of the local pool for locally-
running rust actions when goma is enabled, but RBE is not.

Change-Id: I76bd39f2a081b288383a92ab535ce27fc8670595
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847016
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>