[fuchsia-repo] Fix flakes by tracking removed manifests
Previously we didn't observe manifests that we were no longer interested
in watching, so we needed to poll for a period of time to see if no
events were received. This changes the PackageManifestWatcher to now
include the manifests we stopped watching, so the tests can check we
correctly stopped watching these manifests.

Bug: 124412
Change-Id: I00747de6ebcf039c6ea355724f4b9fac810a6dc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848066
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>