[netstack3] Don't pad outgoing ethernet frames
For ethernet devices, don't pad ethernet frames that are smaller than 60
bytes (without the frame check sequence).

This is a no-op with from the perspective of netstack clients since the
net-device client library already pads buffers to the length requested
by the device. Core and integration tests are updated to accommodate the
new behavior.

Bug: 126190
Change-Id: If6ece8467fa37a6fdab28328afe7bf351b65c1d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835717
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>