[netstack] Fix memory leak when CreateNIC fails
Fix a memory leak where if `stack.CreateNICWithOptions` fails, the
qdisc that was initialized and passed as argument was not getting
cleaned up, and the goroutine(s) spawned and the memory allocated
for the queue(s) would leak. This has been fixed by calling `Close`
on the qdisc, which stops the goroutine(s) and allows the queues to
be garbage collected.

Fixed: 126332
Change-Id: If878de52ec31d317e525d5ccd45ebea712e74e8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849196
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>