[virtio-guest] Rework VirtIO ABI to match C++ style guide.
The current ABI definition used by the virtio display driver is a strict
subset of the VirtIO definitions, which are provided in C.

This CL reimplements the ABI definition to match the readability
standards that we've set for our register definitions. This is the
relevant bar, because the ABI is our interface with the virtualized
hardware, similarly to how the registers are our interface with
(usually) physical hardware.

Test: fx test virtio-gpu-test
Test: fx qemu --virtio -g -N -q /usr/bin
Test: fx shell display-tool squares
Change-Id: Ia7dbda86d845f2ba09a46b370e7ff25e80a1c58a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/848358
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>