[07/14][banjo][display] Remove pixel_format from Image.
Now that image_t.pixel_format is not used by drivers nor
display coordinator anymore, it's safe to delete the field.

This is only used in-tree and it's not exposed as part of SDK,
so this change is both API and ABI safe.

Test: fx build
Bug: 126113
Change-Id: Ib0c42d5ad3e7ea0df670512f23bcef14a1e13b61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/843239
Reviewed-by: Victor Costan <costan@google.com>
API-Review: John Bauman <jbauman@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: John Bauman <jbauman@google.com>