[camera] Fix potential race conditions in device component.
A promise scope object should always be the last thing declared in
a class so that it is the first thing destroyed when an instance of
said class is destroyed. Otherwise scheduled tasks within that
scope may try to access class members that have already been
destroyed.

Change-Id: Ia749bc9425456955b6366b09061023993dba4f55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849837
Reviewed-by: John Sasinowski <jsasinowski@google.com>
Commit-Queue: Matt Thiffault <mthiffault@google.com>