[wlan][tool] fix saving network with password only
This fixes the case where a user specifies an SSID and password to save
a network but doesn't specify the security type or credential type. WPA2
is used by default because most users will have WPA2 or WPA3 networks,
and a network saved with WPA2 can be used to connect to WPA2 or WPA3.

Test: added unit test (fx test donut-lib-tests)
Bug: 126081
Change-Id: I16c743cca2f68fa2421733c507aa9c1b4cba1454
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849396
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>