[banjo] Notify the driver runtime of banjo call
This change introduces driver context changes from banjo calls
into the driver runtime environment. This is done by storing the owning
driver in the server side of protocols and interfaces, and then setting
and desetting it in the runtime. This is an optional feature that can be
enabled by passing true into the *Protocol/*Interface template. This
is also working for async callbacks. ddk-callbacks are not supported.

Bug: 116791
Tests: Manually tested on a sherlock by booting it and running all the
tests under //src/devices:tests

Change-Id: Ic65394794236e81130416c1c4c985ed20894ccb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797946
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>