[ffx][emu]Making the instance name default configurable
This makes the default emulator instance name configurable by setting
`emu.name`. The intent of this feature is to allow build systems or
similar, to set the configuration based on the product/board being run
in order to make it more clear what "type" of emulator is running, and
make it easier to serve the correct packages to the target instance.

Bug: 125893
Change-Id: I7b961d30ae7b5aea8af4ee3e51197157d02836ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849600
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>