[flatland] Ignore rows that may be invalid on sherlock / vim3.
On sherlock and vim3 sometimes the last few rows (on sherlock it's the
last 5 rows, on VIM3 with 1920-width display it's the last 2 rows) can
be missing from the captured bytes.

This change modifies the comparison code so that it ignores rows that
can be invalid.

Bug: 125394
Bug: 125842
Change-Id: I4acb399e1eff913dda7b3781135112e9add7cb66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850120
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>