[fxfs] Add seek table to persistent layer.
Add a seek table to the end of persistent layers that gives the id of
the first part of the search key, allowing for fast approximate search
of the entire layer.

In order to support this id searching method, a trait is added to the
Key trait which documents and requires that each key begin with a u64
that will be the first deciding faster in sorting the overall Key.

Change-Id: I499932f712f2330f6313bf84b99dd6ecd679ef57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/838058
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>