[product] conditionally build product bundle v1
In preparation for removing PBv1, add a flag to avoid building PBv1.
Right now, this defaults to True, so this CL is somewhat of a no-op.
In the future, this will default to False when we believe there
shouldn't be any issue in doing so, but we can still quickly re-enable
it.
In the further future, the PBv1 generation will be removed and these
markers (flag/config checks) will help guide what to delete.

Note: this CL doesn't necessarily remove all the PBv1 functionality.
I.e. I expect to add more checks as issues are found.

Bug: 126216
Change-Id: I5db5b7ec74ec77a6bc3d436b83955f1f25876ee3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/847436
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>