[rbe] download tool command prefix
For commands that might need to use remote outputs that
were not downloaded due to rewrapper --download_outputs=false:

  build/rbe/dlwrap.py --download ... -- command

or

  build/rbe/dlwrap.py --download ... && command

(depending on which is more suitable).

--download arguments are paths that point to places that would
normally would be real files, but could be download stubs.
Real files are ignored and left as-is, stubs (links) are downloaded
and written over the same locations.

Note: this does not require any xattrs because reproxy is not involved.

Bug: 123178
Change-Id: I0351dd07bfb5b34b101a125ed32d58dde6fb74e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851319
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>