[testrunner] Fix nsjail env variable handling
The nsjail wrapper is splitting on `=` and is mutilating environment
variables that include a `=` in the value. Do a strings#Cut instead to
only split on the first `=`, ignore `found` as os.Environ() is
guaranteed to return strings containing the separator.

Bug: 88169
Change-Id: I0f7f7917b05686571701be8fca8e59567cadae75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851416
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>