[radar][proxy] Refactor RadarReaderProxy before burst injection changes
Extract the client instance class out of RadarReaderProxy and narrow the
interface it exposes. This should make each easier class easier to
understand, which will be helpful for the upcoming changes to enabled
burst injection.

Also fix a typo, and make sure the radar client is valid everywhere it
is used.

Bug: 100020
Test: fx test radar-proxy-test
Test: fx test radar-integration-test
Test: Radar features worked on Nelson
Change-Id: I6a84a81e86258bb6ced895c8b98dd70619de2482
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851219
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>