[components][routing] Employ "route bundles"
Routing operations for use, offer, and expose now act on a RouteBundle
type, which embodies a group of one or more routing declarations, either
a singleton (a single route with a unique source) or an aggregate
(multiple routes or one route with a collection source).

This is useful because in some cases it is possible to aggregate multple
capabilities together, and we want the router to have a standard way to
represent this.

Currently, aggregation is implemented by "lookahead" logic that doesn't
mesh well with the of the algorithm. This change makes the handling of
aggregates a regular part of route traversal and reduces the number of
places that need to handle aggregation individually.

This change does not introduce any new forms of aggregation, but without
it, introducing more types of aggregation will be harder. We plan to
base these types of aggregation upon this change:

- Aggregate multiple collections
- Aggregate collections and static children together
- Map an aggregated expose to a single exposed directory entry

Bug: 4776
Change-Id: I9f5c3847a27a78fcc1f8510b4b4daff3637db6e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845021
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Gary Bressler <geb@google.com>