[driver-runner] Sometimes place boot drivers in pkg drivers collection
The placement of drivers based on the package it originates from is too
simple and leads to dependency cycles where a driver in boot-drivers may
end up depending on pkg-drivers. By taking into consideration the
collections of the parents the driver binds to into consideration when
choosing a collection we can ensure that we avoid such dependency loops.

Additionally, the kHost option from the Collection enum was removed as
it was unused.

Bug: 125674
Test: fx test driver-runner-tests
Change-Id: I14ebd318267e659b465c8ec112077f6e7f8dba79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/850436
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>